// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
// ...
      stocks: {
        Row: {
          created_at: string
          departament_id: string
          id: string
          input_quantity: number
          is_active: boolean
          product: string
          updated_at: string
        }
        Insert: {
          created_at?: string
          departament_id: string
          id?: string
          input_quantity: number
          is_active?: boolean
          product: string
          updated_at?: string
        }
        Update: {
          created_at?: string
          departament_id?: string
          id?: string
          input_quantity?: number
          is_active?: boolean
          product?: string
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "stocks_departament_id_fkey"
            columns: ["departament_id"]
            isOneToOne: false
            referencedRelation: "departaments"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}